requirements.txt
python-dotenv==0.21.0
json==2.0.9
argparse==1.1
smtplib==8.8.4